+++ CHANGELOG.md
@@ -4,7 +4,9 @@ All notable changes to Wernanmail will be documented in this file.
 
 ## [Unreleased]
 
-## [0.4.0] - 2026-07-18
+### Added
+
+- Webmail: resume/edit drafts, Archive/Spam/move actions, list previews, load-more pagination, custom folders, Bcc, mark unread.
 
 ### Added